requirements.txt
librosa>=0.10
matplotlib>=3.8
numpy>=1.26
soundfile>=0.12
audioread>=3.0
pillow>=10.0
tqdm